# docs/api/openapi.yaml
openapi: '3.1.0'
info:
  title: Orchestrate OS API
  version: 2.5.0-draft
  description: 'Governance-only semantic control plane API. Resource-style endpoints for multi-user Postgres-backed persistence
    with append-only audit logging.

    '
  contact:
    name: Orchestrate OS Team
servers:
- url: /api/v2.5
  description: Canonical API base
paths:
...
  /documents/{doc_id}/corrections:
    post:
      operationId: createCorrection
      summary: Create a field-level correction
      description: 'Creates a correction linked to a document, optionally referencing an anchor and/or RFI. Auto-classifies
        as minor (auto_applied) or non_trivial (pending_verifier) based on policy: abs(length_delta)<=2 AND no digits AND
        no currency/percent = minor. Emits correction.created audit event. Feature-gated behind EVIDENCE_INSPECTOR_V251.

        '
      tags:
      - Evidence Inspector
      parameters:
      - name: doc_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CorrectionCreate'
      responses:
        '201':
          description: Correction created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CorrectionResponse'
        '400':
          description: Validation error (missing corrected_value, invalid anchor/rfi ref)
        '404':
          description: Document not found or feature disabled
  /corrections/{cor_id}:
    patch:
      operationId: updateCorrection
      summary: Update a correction (approve/reject/modify)
      description: 'Updates correction status, corrected_value, or metadata. Requires version for optimistic concurrency (409
        STALE_VERSION on mismatch). Setting status to approved/rejected auto-sets decided_by and decided_at. Emits correction.updated
        audit event. Feature-gated behind EVIDENCE_INSPECTOR_V251.

        '
      tags:
      - Evidence Inspector
      parameters:
      - name: cor_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CorrectionUpdate'
      responses:
        '200':
          description: Correction updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CorrectionResponse'
        '400':
          description: Validation error
        '404':
          description: Correction not found or feature disabled
        '409':
          description: Stale version — concurrent modification
  /batches/{bat_id}/corrections:
    get:
      operationId: listBatchCorrections
      summary: List corrections across all documents in a batch
      description: 'Returns paginated corrections for all documents in the batch, optionally filtered by status. Feature-gated
        behind EVIDENCE_INSPECTOR_V251.

        '
      tags:
      - Evidence Inspector
      parameters:
      - name: bat_id
        in: path
        required: true
        schema:
          type: string
      - name: status
        in: query
        schema:
          type: string
          enum:
          - pending_verifier
          - approved
          - rejected
          - auto_applied
      - $ref: '#/components/parameters/CursorParam'
      - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Correction list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CorrectionCollection'
        '404':
          description: Batch not found or feature disabled
...
components:
  parameters:
...
    CursorParam:
      name: cursor
      in: query
      required: false
      schema:
        type: string
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
  schemas:
...
    ResponseMeta:
      type: object
      properties:
        request_id:
          type: string
        timestamp:
          type: string
          format: date-time
    CollectionMeta:
      type: object
      properties:
        request_id:
          type: string
        timestamp:
          type: string
          format: date-time
        pagination:
          type: object
          properties:
            cursor:
              type: string
            has_more:
              type: boolean
            limit:
              type: integer
...
    Correction:
      type: object
      properties:
        id:
          type: string
          description: ULID with cor_ prefix
        document_id:
          type: string
        workspace_id:
          type: string
        anchor_id:
          type:
          - string
          - 'null'
        rfi_id:
          type:
          - string
          - 'null'
        field_id:
          type: string
        field_key:
          type: string
        original_value:
          type:
          - string
          - 'null'
        corrected_value:
          type: string
        correction_type:
          type: string
          enum:
          - minor
          - non_trivial
          description: 'Auto-classified: minor if abs(length_delta)<=2, no digits, no currency/percent symbols. Otherwise
            non_trivial.

            '
        status:
          type: string
          enum:
          - pending_verifier
          - approved
          - rejected
          - auto_applied
        decided_by:
          type:
          - string
          - 'null'
        decided_at:
          type:
          - string
          - 'null'
          format: date-time
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deleted_at:
          type:
          - string
          - 'null'
          format: date-time
        version:
          type: integer
        metadata:
          type: object
    CorrectionCreate:
      type: object
      required:
      - corrected_value
      properties:
        original_value:
          type: string
        corrected_value:
          type: string
        anchor_id:
          type: string
        rfi_id:
          type: string
        field_id:
          type: string
        field_key:
          type: string
        metadata:
          type: object
    CorrectionUpdate:
      type: object
      required:
      - version
      properties:
        version:
          type: integer
          description: Required for optimistic concurrency
        status:
          type: string
          enum:
          - pending_verifier
          - approved
          - rejected
          - auto_applied
        corrected_value:
          type: string
        metadata:
          type: object
    CorrectionResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Correction'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    CorrectionCollection:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Correction'
        meta:
          $ref: '#/components/schemas/CollectionMeta'
...
  securitySchemes:
    GoogleOAuth:
      type: oauth2
      description: 'Google OAuth 2.0 (OIDC) for production human users. Required for human-governed endpoints (patch transitions,
        approvals, annotations).

        '
      flows:
        authorizationCode:
          authorizationUrl: https://accounts.google.com/o/oauth2/v2/auth
          tokenUrl: https://oauth2.googleapis.com/token
          scopes:
            email: Read user email
            profile: Read user profile
    BearerToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Server-issued JWT (1h expiry) obtained after Google OAuth code exchange. Used in Authorization header
        for all authenticated requests.

        '
    ScopedApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: 'Workspace-scoped API key for sandbox and service ingestion endpoints. Keys carry explicit scopes (e.g.,
        signals:write, batches:write, read:all). Required for service ingestion endpoints (POST /signals, POST /triage-items).
        Accepted as alternative auth on read endpoints.

        '
security:
- BearerToken: []
- ScopedApiKey: []